refactoring and splitting
