+++ cat-app-mnb/src/components/modal/Modal.js
@@ -1,8 +1,6 @@
 import classes from "./Modal.module.css";
 
 const Modal = (props) => {
-  console.log("props from modal", props.modData);
-
   const catData = props.modData;
 
   return (
@@ -14,11 +12,7 @@ const Modal = (props) => {
           </div>
           <div className={classes.title}></div>
 
-          <img
-            className={classes.modalCatImg}
-            src={catData.url}
-            alt=""
-          ></img>
+          <img className={classes.modalCatImg} src={catData.url} alt=""></img>
           <div className={classes.wrapper}>
             <p>
               Cat URL:
