+++ cat-app-mnb/src/pages/AllCats.module.css
@@ -1,4 +1,3 @@
-/* HOW to center the button in the main page */
 .mainContainer {
   display: flex;
   flex-direction: column;
@@ -6,7 +5,7 @@
   align-items: center;
 }
 
-.openModalBtn {
+.loadMoreBtn {
   width: 200px;
   height: 40px;
   border: none;
@@ -15,8 +14,3 @@
   color: white;
   cursor: pointer;
 }
-
-/* .body {
-  padding: 0;
-  margin: 0;
-} */
